Seperated test case categories to their own files
