+++ .vscode/settings.json
@@ -38,5 +38,6 @@
         "qualifyMembers",
         "addOverride"
     ],
-    "java.debug.settings.vmArgs": "-ea"
+    "java.debug.settings.vmArgs": "-ea",
+    "java.debug.settings.onBuildFailureProceed": true
 }
